requirements.txt
Django==1.9.7
django-allauth==0.27.0
django-cities==0.4.2
django-model-utils==2.6
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==2.0.0
Pillow==3.4.1
psycopg2==2.6.2
python-openid==2.2.5
qrcode==5.3
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
wheel==0.24.0
git+git://github.com/ojii/pymaging.git#egg=pymaging
git+git://github.com/ojii/pymaging-png.git#egg=pymaging-png